+++ dependencies.gradle
@@ -4,7 +4,7 @@
 // also being more explicit about what is depended on.
 
 dependencies {
-    api('com.github.GTNewHorizons:Steve-s-Factory-Manager:1.3.0-GTNH:dev')
+    api('com.github.GTNewHorizons:Steve-s-Factory-Manager:1.2.0-GTNH:dev')
 
     api('curse.maven:cofh-core-69162:2388751')
